Update messaging.pem
Remove the *.messaging.internetofthings.ibmcloud.com certificate from the .pem in preparation for renewal of the *.messaging.internetofthings.ibmcloud.com certificate in public instances of the platform. Removal from the .pem should not affect the ability to connect as the trust chain for the renewed certificate presented by Watson IoT Platform will remain the same.